Removed unseeable screenshots :(
Forge never added a way to view screenshots in the mod menu despite mcmod.info having a section to put your screenshots in. I've moved these to the docs folder for now, so that they aren’t bundled with the mod on build. The upshot of this is that the jar's size is greatly reduced (the JavaDoc jar has a larger file size than the mod right now!), but the downside is that nobody gets a fun extra :(