[WIP] Non-multi threaded ender bow, part 3
When saving and re-loading a world, you don't loose your arrows now! This took too long to code, but this was mostly because I wrote a very over engineered system to save and load an arbitrary amount of entities. This is still a very WIP bit of code, and I think this now introduces more bugs around arrow velocities not being displayed properly client side. On to fixing that next...