More WIP code cleanups
Most of these "optimisations" are to reduce compiled .class file size. My target is to keep the .jar builds under 64K, which is currently being met! The good thing is that reducing the size of the program is helping to reduce the complexity of the program as well, although I really need to document why I've made certain decisions.