WIP custom bow renderer progress
This code is awful!
- The first person "draw back" animation mostly works, but the bow flexes in a very strange way while drawing it back. I can’t exactly figure out why, I'll continue searching for a fix...
- Bows now also render correctly on non-player entities! I've even made them render correctly for witches, not that there’s any reason to.

Unfortunately, the .jar file size is now above 64k! This is mostly due to the horrendous amount of code I had to copy & paste to get the custom renderer working, plus the fact that the custom renderer needs a new class. Curse you, Forge 1.7.10! If only we had mixins back in the day, I'd be able to inject a couple of lines of code into the checks against the bow, and I wouldn’t have to do any of this! I understand this file size is completely unacceptable, and I will work on making it smaller...