Optimise mod logo by cropping whitespace, clean up and compress CurseForge screenshot
The mod logo was cropped by ImageMagick's trim operator, which is designed to only remove the border around an image without affecting any other image data. This means that only "whitespace" (purely transparent pixels) was removed from the image. This was tested and confirmed to be true by using various commands to resize the cropped image back to its original size, and then comparing it to the original image with the AE metric (which reported that both images were identical). The reason the image cropped to 541x99 is that there are some stray pixels around the sides of the image, so removing them would remove non-whitespace image data. In a purely pedantic sense, the trim operation is technically lossy, as the original image dimensions are lost, but no image data that isn't purely transparent has been removed, and the original image can be reconstructed from the cropped version with no loss of data.

I've also compressed the CurseForge screenshot in a lossy way, by heavily reducing the colorspace of the image, and cropping it. Dithering was used to help simulate a larger amount of colors. I re-took the screenshot without any sub-pixel rendering, as I find it distracting and messy, and it adds completely superflous data to the image.