Update README to use image links from GitHub pages branch, add "responsive images"
The images don’t respond to you, but they do load in different sizes now! I don’t like using the hardcoded links, but it seems to be the best way to load the images regardless of where you are (i.e. it works when I try to preview markdown while locally working on the README).