Use lossy compression techniques to reduce size of logo included in compiled .jar file
- All semi-transparent shadows have been removed from the logo via ImageMagick, as the semi-transparent shadows were hard to see against the mod menu background.
- The amount of colours in the logo has been reduced using the version of pngquant included in ImageAlpha (https://github.com/kornelski/ImageAlpha).
- The resulting .png file was then optimised by all the usual programs I use to optimise .png files (detailed in previous commits).

I was pretty happy with how this turned out, as I can barely see a difference when looking at the logo in the mod menu! The image file now takes up only 18% of the space it used to, which is very good considering that it previously took up more than half of the space of the compiled mod .jar.